actualizacion de todo

+++ Backend/fashionStore_back/API.Data/Entidades/Gestion/Nomencladores/Producto.cs
@@ -4,8 +4,8 @@
     {
         public required string Codigo { get; set; }
         public required string Descripcion { get; set; }
-        public required float Costo { get; set; }
-        public required float PrecioUSD { get; set; }
+        public required decimal Costo { get; set; }
+        public required decimal PrecioUSD { get; set; }
  
         public required int Cantidad { get; set; }
         public required int CantidadStock { get; set; }
